requirements.txt
Django==4.1.6
python-telegram-bot[callback-data]==20.1
requests
python-decouple==3.7